requirements.txt
langchain-google-genai
langchain-core
neo4j
python-dotenv
pydantic
